Fix: Make app images optional and improve rendering
The `image` and `imageProviderFileId` columns in the `app` table have
been made optional. This allows apps to be created without images.

Additionally, the image rendering logic in several components has been
updated to handle cases where an image is not present, displaying a
placeholder instead. The `sizes` attribute has also been added to
`Image` components for better performance.

The padding on comments has been reduced from 6 to 1 for better spacing.